Add files via upload
Project description.
This project is about preparing a report for a bank’s loan division. Step by step I explain how to find out if a customer’s marital status and number of children have an impact on whether they will default on a loan. The bank already has some data on customers’ creditworthiness.
I build a credit score for a potential customer. A credit score is used to evaluate the ability of a potential borrower to repay their loan.